added magnetic field and rotation vector listeners, added signal classes, changed sensor data listener method, changed buffering method, changed saving method, GUI changes,

+++ app/src/main/java/com/example/divided/signalgathering/ChartPoint.java
@@ -17,7 +17,7 @@ public class ChartPoint {
         return timeStamp;
     }
 
-    public void setTimeStamp(float timeStamp){
+    public void setTimeStamp(long timeStamp){
         this.timeStamp = timeStamp;
     }
 }
